pyproject.toml
[tool.poetry]
name = "fastapi-rag-backend"
version = "0.1.0"
description = "FastAPI-based RAG backend for Korean tourism chatbot"
authors = ["Your Name <your.email@example.com>"]
readme = "readme.md"
python = "^3.9"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
langchain = "~0.2.0"
langchain-openai = "^0.1.0"
openai = "^1.12.0,<2.0.0"
pgvector = "^0.2.4"
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
black = "^24.0.0"
ruff = "^0.1.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--maxfail=1 --disable-warnings -v"
asyncio_mode = "auto"
